More data model work

+++ Library/C.framework/Headers/stdint.h
@@ -124,17 +124,17 @@
 
 typedef signed char         int8_t;
 typedef signed short        int16_t;
-typedef signed int          int32_t;
+typedef signed long         int32_t;
 typedef signed long long    int64_t;
 
 typedef signed short        int_fast8_t;
 typedef signed short        int_fast16_t;
-typedef signed int          int_fast32_t;
+typedef signed long         int_fast32_t;
 typedef signed long long    int_fast64_t;
 
 typedef signed char         int_least8_t;
 typedef signed short        int_least16_t;
-typedef signed int          int_least32_t;
+typedef signed long         int_least32_t;
 typedef signed long long    int_least64_t;
 
 
@@ -144,17 +144,17 @@ typedef __intptr_t intptr_t;
 
 typedef unsigned char       uint8_t;
 typedef unsigned short      uint16_t;
-typedef unsigned int        uint32_t;
+typedef unsigned long       uint32_t;
 typedef unsigned long long  uint64_t;
 
 typedef unsigned short      uint_fast8_t;
 typedef unsigned short      uint_fast16_t;
-typedef unsigned int        uint_fast32_t;
+typedef unsigned long       uint_fast32_t;
 typedef unsigned long long  uint_fast64_t;
 
 typedef unsigned char       uint_least8_t;
 typedef unsigned short      uint_least16_t;
-typedef unsigned int        uint_least32_t;
+typedef unsigned long       uint_least32_t;
 typedef unsigned long long  uint_least64_t;
